+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/AnnotationHandlerMapping.java
@@ -1,15 +1,12 @@
 package com.interface21.webmvc.servlet.mvc.tobe;
 
-import com.interface21.context.stereotype.Controller;
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
-import org.reflections.Reflections;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
@@ -17,7 +14,6 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 public class AnnotationHandlerMapping {
 
@@ -32,24 +28,13 @@ public class AnnotationHandlerMapping {
     }
 
     public void initialize() {
-        Reflections reflections = new Reflections(basePackage);
-        Set<Class<?>> controllers = reflections.getTypesAnnotatedWith(Controller.class);
-        for (Class<?> controller : controllers) {
-            Object instance = createInstance(controller);
-            List<Method> methods = getRequestMethods(controller);
-            addHandlers(instance, methods);
+        ControllerScanner controllerScanner = new ControllerScanner(basePackage);
+        for (Map.Entry<Class<?>, Object> entry : controllerScanner.controllers.entrySet()) {
+            List<Method> methods = getRequestMethods(entry.getKey());
+            addHandlers(entry.getValue(), methods);
         }
 
         log.info("Initialized AnnotationHandlerMapping!");
-    }
-
-    private Object createInstance(Class<?> clazz) {
-        try {
-            Constructor<?> constructor = clazz.getConstructor();
-            return constructor.newInstance();
-        } catch (NoSuchMethodException | InvocationTargetException | InstantiationException | IllegalAccessException e) {
-            throw new RuntimeException("Can't create instance of class : " + clazz.getName(), e);
-        }
     }
 
     private List<Method> getRequestMethods(Class<?> clazz) {
